Drop the remaining float casts on pi
vSunlight only ever reaches the GL renderer as a sun direction and a shading dot product, so the eighth decimal
place there buys nothing. The two in SpawnEncounterMonsters place random encounter monsters, and no trace in
test/Data reaches that function - instrumenting it and replaying the whole suite gives zero hits - so no
recorded trace can desync on this.

M_PI is now spelled the same way everywhere.

+++ src/Engine/Graphics/Indoor.cpp
@@ -1754,7 +1754,7 @@ int SpawnEncounterMonsters(MapInfo *map_info, int enc_index) {
         for (; loop_cnt < 100; ++loop_cnt) {
             // random x,y at distance from party
             dist_from_party = grng->random(1024) + 512;
-            angle_from_party = (grng->random(TrigLUT.uIntegerDoublePi) * 2 * static_cast<float>(M_PI)) / TrigLUT.uIntegerDoublePi;
+            angle_from_party = (grng->random(TrigLUT.uIntegerDoublePi) * 2 * M_PI) / TrigLUT.uIntegerDoublePi;
             enc_spawn_point.position.x = pParty->pos.x + std::cos(angle_from_party) * dist_from_party;
             enc_spawn_point.position.y = pParty->pos.y + std::sin(angle_from_party) * dist_from_party;
             enc_spawn_point.position.z = pParty->pos.z;
@@ -1790,7 +1790,7 @@ int SpawnEncounterMonsters(MapInfo *map_info, int enc_index) {
         for (loop_cnt = 0; loop_cnt < 100; ++loop_cnt) {
             // random x,y at distance from party
             dist_from_party = grng->random(512) + 256;
-            angle_from_party = (grng->random(TrigLUT.uIntegerDoublePi) * 2 * static_cast<float>(M_PI)) / TrigLUT.uIntegerDoublePi;
+            angle_from_party = (grng->random(TrigLUT.uIntegerDoublePi) * 2 * M_PI) / TrigLUT.uIntegerDoublePi;
             enc_spawn_point.position.x = pParty->pos.x + std::cos(angle_from_party) * dist_from_party;
             enc_spawn_point.position.y = pParty->pos.y + std::sin(angle_from_party) * dist_from_party;
             enc_spawn_point.position.z = pParty->pos.z;
